♻️ refactor(ui): 폴더 공유 바텀 시트 구조 개선 및 기능 고도화
- `_ShareBottomSheet.kt` 파일을 삭제하고 `ShareBottomSheet.kt`로 로직 통합 및 전면 재작성
- 공유 프로세스의 단계적 상태 관리를 위해 `SelectDepth`, `ScreenState`, `LinkGenerateState` 도입
- 폴더 시각화 컴포저블(`FolderStructure`) 및 카테고리/폴더 탐색 UI(`SelectFolderToShareScreen`) 구현
- 링크 생성 상태(Before, Loading, Done, Error)에 따른 하단 버튼 분기 및 스켈레톤 로직 반영
- 시스템 클립보드 복사 및 안드로이드 Share Sheet(`Intent.ACTION_SEND`) 연동 로직 추가
- 디자인 가이드에 맞춘 레이아웃 비율 상수 적용 및 UI 컴포넌트 세분화
- 컴포넌트 내 ViewModel 직접 참조를 제거하고 상태와 콜백을 주입받는 구조로 개선하여 재사용성 확보
